Fix minor bugs

+++ Benchmark/ValidateTFModel.py
@@ -7,7 +7,7 @@ interpreter.allocate_tensors()
 input_details = interpreter.get_input_details()
 output_details = interpreter.get_output_details()
 
-dummy_input = np.zeros(input_details[0]['shape'], dtype=np.int16)
+dummy_input = np.zeros(input_details[0]['shape'], dtype=np.int8)
 interpreter.set_tensor(input_details[0]['index'], dummy_input)
 
 interpreter.invoke()
